+++ src/query/service/Cargo.toml
@@ -64,8 +64,10 @@ common-storages-factory = { path = "../storages/factory" }
 common-storages-fuse = { path = "../storages/fuse" }
 common-storages-hive = { path = "../storages/hive", optional = true }
 common-storages-index = { path = "../storages/index" }
+common-storages-information-schema = { path = "../storages/information-schema" }
 common-storages-preludes = { path = "../storages/preludes" }
 common-storages-share = { path = "../storages/share" }
+common-storages-view = { path = "../storages/view" }
 common-streams = { path = "../streams" }
 common-tracing = { path = "../../common/tracing" }
 common-users = { path = "../users" }
