+++ src/query/service/src/interpreters/interpreter_view_alter.rs
@@ -22,12 +22,12 @@ use common_meta_app::schema::DropTableReq;
 use common_meta_app::schema::TableMeta;
 use common_meta_app::schema::TableNameIdent;
 use common_planner::plans::AlterViewPlan;
+use common_storages_view::view_table::VIEW_ENGINE;
 
 use crate::interpreters::Interpreter;
 use crate::pipelines::PipelineBuildResult;
 use crate::sessions::QueryContext;
 use crate::sessions::TableContext;
-use crate::storages::view::view_table::VIEW_ENGINE;
 
 pub struct AlterViewInterpreter {
     ctx: Arc<QueryContext>,
