+++ src/query/storages/information-schema/Cargo.toml
@@ -1,10 +1,10 @@
 [package]
-name = "common-storages-factory"
-version = "0.1.0"
-edition = "2021"
-authors = ["Databend Authors <opensource@datafuselabs.com>"]
-license = "Apache-2.0"
-publish = false
+name = "common-storages-information-schema"
+version = { workspace = true }
+authors = { workspace = true }
+license = { workspace = true }
+publish = { workspace = true }
+edition = { workspace = true }
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 [lib]
@@ -14,28 +14,39 @@ test = false
 [dependencies]
 common-base = { path = "../../../common/base" }
 common-catalog = { path = "../../catalog" }
-common-config = { path = "../../config" }
 common-datablocks = { path = "../../datablocks" }
 common-datavalues = { path = "../../datavalues" }
 common-exception = { path = "../../../common/exception" }
-common-fuse-meta = { path = "../fuse-meta" }
+common-formats = { path = "../../formats" }
+common-functions = { path = "../../functions" }
 common-meta-app = { path = "../../../meta/app" }
 common-meta-types = { path = "../../../meta/types" }
+common-metrics = { path = "../../../common/metrics" }
 common-pipeline-core = { path = "../../pipeline/core" }
+common-pipeline-sinks = { path = "../../pipeline/sinks" }
+common-pipeline-sources = { path = "../../pipeline/sources" }
 common-pipeline-transforms = { path = "../../pipeline/transforms" }
 common-planner = { path = "../../planner" }
-common-storages-fuse = { path = "../fuse" }
-common-storages-index = { path = "../index" }
-common-storages-preludes = { path = "../preludes" }
+common-storage = { path = "../../../common/storage" }
 common-storages-view = { path = "../view" }
-common-streams = { path = "../../streams" }
+common-users = { path = "../../users" }
 
 async-trait = { version = "0.1.57", package = "async-trait-fn" }
 backon = "0.2"
+chrono = "0.4.22"
 futures = "0.3.24"
+itertools = "0.10.5"
+once_cell = "1.15.0"
 opendal = { version = "0.19", features = ["layers-retry", "layers-tracing", "layers-metrics", "compress"] }
 parking_lot = "0.12.1"
 serde = { workspace = true }
 serde_json = { workspace = true }
+serde_repr = "0.1.9"
+snailquote = "0.3.1"
 tracing = "0.1.36"
+typetag = "0.2.3"
 uuid = { version = "1.1.2", features = ["serde", "v4"] }
+walkdir = "2.3.2"
+
+[build-dependencies]
+common-building = { path = "../../../common/building" }
