+++ src/query/storages/information-schema/src/lib.rs
@@ -11,7 +11,6 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
-
 mod columns_table;
 mod keywords_table;
 mod schemata_table;
