+++ src/query/storages/view/Cargo.toml
@@ -1,0 +1,21 @@
+[package]
+name = "common-storages-view"
+version = { workspace = true }
+authors = { workspace = true }
+license = { workspace = true }
+publish = { workspace = true }
+edition = { workspace = true }
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+[lib]
+doctest = false
+test = false
+
+[dependencies]
+common-catalog = { path = "../../catalog" }
+common-exception = { path = "../../../common/exception" }
+common-meta-app = { path = "../../../meta/app" }
+
+async-trait = { version = "0.1.57", package = "async-trait-fn" }
+
+[build-dependencies]
