+++ src/query/storages/view/src/lib.rs
@@ -1,4 +1,4 @@
-// Copyright 2021 Datafuse Labs.
+// Copyright 2022 Datafuse Labs.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -13,4 +13,3 @@
 // limitations under the License.
 
 pub mod view_table;
-pub use view_table::ViewTable;
