+++ src/query/storages/view/src/view_table.rs
@@ -15,12 +15,11 @@
 use std::any::Any;
 use std::sync::Arc;
 
+use common_catalog::catalog::StorageDescription;
+use common_catalog::table::Table;
 use common_exception::ErrorCode;
 use common_exception::Result;
 use common_meta_app::schema::TableInfo;
-
-use crate::storages::StorageDescription;
-use crate::storages::Table;
 
 pub struct ViewTable {
     table_info: TableInfo,
